Merges dev into feat/FE-4-help-page
